refactor(spec): split tool search execution surfaces
